+++ core/shared/src/main/scala/zio/ZIO.scala
@@ -1527,7 +1527,7 @@ private[zio] trait ZIOFunctions extends Serializable {
    *
    * Unlike `foreachAllPar`, this method will use at most `n` fibers.
    */
-  final def collectAllParN[R, E, A](n: Long)(as: Iterable[ZIO[R, E, A]]): ZIO[R, E, List[A]] =
+  final def collectAllParN[R, E, A](n: Int)(as: Iterable[ZIO[R, E, A]]): ZIO[R, E, List[A]] =
     foreachParN[R, E, ZIO[R, E, A], A](n)(as)(ZIO.identityFn)
 
   /**
@@ -1768,7 +1768,7 @@ private[zio] trait ZIOFunctions extends Serializable {
    * Unlike `foreachPar`, this method will use at most up to `n` fibers.
    */
   final def foreachParN[R, E, A, B](
-    n: Long
+    n: Int
   )(as: Iterable[A])(fn: A => ZIO[R, E, B]): ZIO[R, E, List[B]] =
     Queue
       .bounded[(Promise[E, B], A)](n.toInt)
@@ -1791,10 +1791,10 @@ private[zio] trait ZIOFunctions extends Serializable {
    * Unlike `foreachPar_`, this method will use at most up to `n` fibers.
    */
   final def foreachParN_[R, E, A](
-    n: Long
+    n: Int
   )(as: Iterable[A])(f: A => ZIO[R, E, _]): ZIO[R, E, Unit] =
     Semaphore
-      .make(n)
+      .make(n.toLong)
       .flatMap { semaphore =>
         ZIO.foreachPar_(as) { a =>
           semaphore.withPermit(f(a))
@@ -2080,7 +2080,7 @@ private[zio] trait ZIOFunctions extends Serializable {
   /**
    *  Alias for [[ZIO.collectAllParN]]
    */
-  final def sequenceParN[R, E, A](n: Long)(as: Iterable[ZIO[R, E, A]]): ZIO[R, E, List[A]] =
+  final def sequenceParN[R, E, A](n: Int)(as: Iterable[ZIO[R, E, A]]): ZIO[R, E, List[A]] =
     collectAllParN[R, E, A](n)(as)
 
   /**
@@ -2182,7 +2182,7 @@ private[zio] trait ZIOFunctions extends Serializable {
    * Alias for [[ZIO.foreachParN]]
    */
   final def traverseParN[R, E, A, B](
-    n: Long
+    n: Int
   )(as: Iterable[A])(fn: A => ZIO[R, E, B]): ZIO[R, E, List[B]] =
     foreachParN[R, E, A, B](n)(as)(fn)
 
@@ -2190,7 +2190,7 @@ private[zio] trait ZIOFunctions extends Serializable {
    * Alias for [[ZIO.foreachParN_]]
    */
   final def traverseParN_[R, E, A](
-    n: Long
+    n: Int
   )(as: Iterable[A])(f: A => ZIO[R, E, _]): ZIO[R, E, Unit] =
     foreachParN_[R, E, A](n)(as)(f)
